First/Final Commit (JavaScript)
This is my first/final commit and is ready for review. In this commit I added in code to declared the variable thisTime containing the date object. Also, the thisHour and thisMonth is created to be able to extract the hour and month value from the thisTime variable. Furthermore, I used the inline element to show the sky image in the web page. Finally, inserted the value of the imgStr variable  directly after the element's opening tag.